requirements.txt
Django==1.4.2
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.18
celery==3.0.12
django-celery==3.0.11
kombu==2.4.8
python-dateutil==1.5
pytz==2012h
wsgiref==0.1.2
-e git+git@github.com:BNOTIONS/stallion.git#egg=stallion
PIL
boto
django_storages